+++ apps/csv/manifest.h
@@ -1,0 +1,11 @@
+/* manifest.h -- the datasets cmuts csv reads.
+ *
+ * Author: Hamish M. Blair <hmblair@stanford.edu>
+ */
+
+#pragma once
+
+#include "format.h"
+
+extern const fmt_request CMUTS_CSV_READS[];
+extern const size_t      CMUTS_CSV_N_READS;