+++ include/subcommands.h
@@ -14,3 +14,4 @@ int sub_main(int argc, char **argv);
 int div_main(int argc, char **argv);
 int norm_main(int argc, char **argv);
 int score_main(int argc, char **argv);
+int csv_main(int argc, char **argv);